Some changes.
message.guild.member(message.mentions.members.first()) | You're basically getting the member object of the first member mentioned, which is redundant in this case. 

auditlogchannel.send(embed) | You should add a catch statement so if the channel is not found, the bot will still ban the member while logging the error.

You should consider using the <GuildMember>.bannable property to check if the member can be banned before actually attempting to do so.