Update kick.js
Same ideas as the ban command. Didn't really go through this one as much but I wanted to point something out.
`await user.send(embed)` | The bot will error if it tries to send a message to a user that either: a) blocked the bot, or b) has privacy settings set to not receive DMs from members of a server. Adding a `.catch(e => { })` will allow the bot to silently let this error go without stopping the kick process.